style: make the workspace clippy-clean under CI's -D warnings
CI runs `cargo clippy --workspace --all-targets -- -D warnings`; newer clippy
(1.95) flags lints the code predates. Bring the whole workspace to zero
warnings/errors with no behavior changes:

- Auto-applied machine-fixable lints workspace-wide (unnecessary_cast,
  manual_range_contains, manual_is_multiple_of, redundant_pattern_matching,
  needless_return, question_mark, map_identity, useless_conversion,
  unnecessary_sort_by -> sort_by_key, ...).
- Removed ~40 redundant `let x = x;` rebindings in dispatch/{fs,net}.rs.
- Promoted constant `assert!`s to `const { assert!(...) }` (compile-time).
- Targeted, justified `#[allow]`s: genuinely-wide signatures (inject_signal),
  parallel-array index loops (page_table), a sharded-map field type, a
  once-per-run enum, and the duplicate-mod test-harness pattern.
- doc-comment list formatting (de-indented markers + paragraph separators).
- Added `OwnedHostMapping::is_empty`; dropped a dead `NOFILE_HARD` const (the
  live copy is a local const in dispatch/time.rs).

No-panic gate (unwrap/expect/panic denied) resolved without weakening it:
- carrick-image layer ordering returns the existing `OciBootstrapError`
  instead of `expect()` on a provably-non-empty bucket.
- fifo_beacon `STATE.lock().unwrap()` -> `.unwrap_or_else(|e| e.into_inner())`
  (recover from poison rather than panic).
- inotify: `let Some(record) = pop_front() else { break }` (front() already
  proved Some); fcntl F_SETOWN_EX uses explicit byte arrays (no try_into
  unwrap); pathcodec uses `unwrap_or(REPLACEMENT_CHARACTER)` for a
  provably-valid scalar.
- Intentionally-panicking test helpers (not `#[test]` fns, so
  allow-*-in-tests doesn't reach them) get a targeted `#[allow]`.

Verified: clippy --all-targets -D warnings, cargo fmt --check, cargo build
--workspace, cargo test --workspace --lib, and the consolidated integration
suite all green.

+++ crates/carrick-hvf/src/darwin_kqueue.rs
@@ -147,7 +147,7 @@ impl Kevent {
         Self::new(
             pid as usize,
             libc::EVFILT_PROC,
-            (libc::EV_ADD | libc::EV_ONESHOT) as u16,
+            libc::EV_ADD | libc::EV_ONESHOT,
             libc::NOTE_EXIT | libc::NOTE_EXITSTATUS,
         )
     }
@@ -156,7 +156,7 @@ impl Kevent {
     /// `proc_exit` is `EV_ONESHOT` (auto-removed once it fires), so this is only
     /// needed to drop a watch whose wait was interrupted before the exit fired.
     pub fn proc_exit_delete(pid: i32) -> Self {
-        Self::new(pid as usize, libc::EVFILT_PROC, libc::EV_DELETE as u16, 0)
+        Self::new(pid as usize, libc::EVFILT_PROC, libc::EV_DELETE, 0)
     }
 
     /// Stash a small integer (a guest fd) in `udata` so a returned event maps
@@ -236,14 +236,14 @@ impl Kevent {
         Self::new(
             fd as usize,
             libc::EVFILT_VNODE,
-            (libc::EV_ADD | libc::EV_CLEAR) as u16,
+            libc::EV_ADD | libc::EV_CLEAR,
             note,
         )
     }
 
     /// Remove a previously-added `EVFILT_VNODE` watch for `fd`.
     pub fn vnode_delete(fd: RawFd) -> Self {
-        Self::new(fd as usize, libc::EVFILT_VNODE, libc::EV_DELETE as u16, 0)
+        Self::new(fd as usize, libc::EVFILT_VNODE, libc::EV_DELETE, 0)
     }
 
     /// The fd a returned `EVFILT_VNODE` event refers to (its `ident`).
